+++ swing/common-model/src/main/java/module-info.java
@@ -25,6 +25,7 @@
  * <li>{@link is.codion.swing.common.model.worker.ProgressWorker}
  * </ul>
  */
+@org.jspecify.annotations.NullMarked
 module is.codion.swing.common.model {
 	requires transitive java.desktop;
 	requires transitive is.codion.common.model;